Create LeetCode16th.c
Topic introduction：Given an array of N integers, i.e., "nums" and a target value, i.e., "target". Find the three integers in "num s" so that their sum is closest to "target". Returns the sum of these three numbers. Assume that there is only one answer for each input.

 reflections：I have a deeper understanding of double pointers. I think double pointers are usually used for faster traversal of arrays. Usually, the first pointer is at the head of the array and the last pointer is at the tail of the array. The movement of pointers on both sides is often regular. For example, in question 11, we need to calculate the maximum volume, so we have to remove the smallest board on both sides every time we move The precondition is that the array is ordered. When the number is too large, we need to make the number smaller. Naturally, we should let the tail pointer move to the right, and vice versa.